content-security-policy/script-src/script-src-strict_dynamic_double_policy_report_only.html

<!DOCTYPE HTML>
<html>

<head>
    <title>A separate Report-Only policy does not influence `strict-dynamic` in the script-src directive.</title>
    <script src='/resources/testharness.js' nonce='dummy'></script>
    <script src='/resources/testharnessreport.js' nonce='dummy'></script>

    <!-- CSP served:
    1) Content-Security-Policy: script-src 'strict-dynamic' 'nonce-dummy'
    2) Content-Security-Policy-Report-Only: script-src 'none'
    -->
</head>

<body>
    <h1>A separate Report-Only policy does not influence `strict-dynamic` in the script-src directive.</h1>
    <div id='log'></div>

    <script nonce='dummy'>
        async_test(function(t) {
            var handler = function(e) {
                assert_equals(e.violatedDirective.split(' ')[0], 'script-src');
                // Check that every violation comes from the Report-Only policy.
                assert_equals(e.originalPolicy.trim(), "script-src 'none'");
            };
            document.addEventListener('securitypolicyviolation', handler);
            var e = document.createElement('script');
            e.id = 'appendChild-reportOnly';
            e.src = 'simpleSourcedScript.js';
            e.onload = t.step_func_done(function() {
                assert_equals(document.getElementById('appendChild-reportOnly').getAttribute('executed'), '1');
            });
            e.onerror = t.unreached_func('Error should not be triggered.');
            document.body.appendChild(e);
        }, 'Script injected via `appendChild` is allowed with `strict-dynamic` + Report-Only `script-src \'none\'` policy.');
    </script>
</body>

</html>
